refector the Formfiled and added new component called InputField
