+++ src/ui.js
@@ -1,18 +1,19 @@
 const colors = {
-    gray1: "#333333",
-    pink1: "#FFA494",
-    pink2: "#FEC5BB",
-    pink3: "#FAE1DD",
-    pink4: "#FFE5D9",
-    orange:"#FFA85F",
-    orange_ligth:"#FEC89A",
-    blue: "#204186",
-    blue_ligth: "#5E81F4",
-    sky_blue: "#DFEAFF",
-    gray:"#7284A3",
-    gray_ligth:"#CCCCCC",
-    gray_ligth2:"#A39BA8",
-    white:"#F8EDEB",
-  };
+  gray1: "#333333",
+  pink1: "#FFA494",
+  pink2: "#FEC5BB",
+  pink3: "#FAE1DD",
+  pink4: "#FFE5D9",
+  orange: "#FFA85F",
+  orange_ligth: "#FEC89A",
+  black: "#455064",
+  blue: "#204186",
+  blue_ligth: "#5E81F4",
+  sky_blue: "#DFEAFF",
+  gray: "#7284A3",
+  gray_ligth: "#CCCCCC",
+  gray_ligth2: "#A39BA8",
+  white: "#F8EDEB",
+};
 
-  export { colors };
+export { colors };
